DX improvement: Added the -d flag to 'yarn postgres'.
I always use the -d flag and I just think it makes sense
to keep it there. I mean why not?

The -d flag runs Docker in 'detached' mode. Without -d,
docker attaches the command to your terminal, so you have
to open up another terminal to do more stuff, and if you close
the terminal then Docker will be terminated.

With -d, it runs the db in the background.