Instantiate ApolloServer in a context builder fn
To ensure only one instance of knex will be used for all db requests, a
context builder function will be used to instantiate knex, daos,
services, resolvers, schema, and server in a chain and return a new
context which includes the Apollo Server. This function will be called
only once in index.ts.

A buildTestServerContext can be made later where mocked daos are passed
in for unit tests.