Extend User type to include posts in Post's typedef
Previously, the User typeDef in user.ts had a posts field. This is
not the best design because user.ts would need to be continually
modified to include more fields as we add more entities like
Projects and Comments, increasing coupling.

A better pattern is to extend the User type in Post's typeDef. This way
everything related to posts is kept in its own typeDef, even those
that modify the User type. The User typeDef in user.ts can remain
oblivious to Posts, reducing coupling. If we later remove Posts
from our schema, we won't get an error in user.ts.