Add silent compilation of .ts files in /test
We want to check for TypeScript errors in .ts files in the test folder
as part of the CI pipeline. Thus, I configured the tsconfig in /test to
not output when compiling, and configured the build script, which runs
as part of the pipeline, to compile both the "src" and "test" folders.