Mock Express's 'res' object in tests
buildServerContext passes on the `req` and `res` from Express when
ApolloServer is running on http, but the test server does not run
on http and thus cannot use the 'req' and 'res' from Express. Thus,
during testing, the server will crash when calling `res.cookie` to
set the cookie.

I created a mocked context for the test server where `res.cookie`
is mocked. Thus, the test server no longer crashes when trying to set
the cookie.