Modify ServerContext interface to only contain what we need
Previously, ServerContext contained Express' `res` and `req`, which were
passed down to the resolvers. This is not good design as those
objects are more than what we needed, as we were only using res.cookie
to set the cookie. This also couples the backend to the Express
framework and its `res` and `req` API.

I have thus modified ServerContext to include only a setCookie function,
which is all we need for now. Later, things like userId and maybe
authScope will be part of the ServerContext, instead of sending down req
and processing it manually on the resolver side.