Reimplement TestManager's `query` method using supertest
Since we are working with supertest, which is a bit lower-level than the
previous testClient hook from Apollo-Server-Testing, the query method
now takes in the GraphQL DocumentNode directly, converted to string
using graphql's `print` method.

The raw response now includes http headers, and the actual data is now
in a JSON string, so we must parse it to achieve what testClient was
previously doing for us.