Add DateTime custom scalar to timestamp fields
Previously, timestamps were typedef'd as string and timestamps
were returned as unix time in ms. Adding this custom scalar allows
automatic serialization of timestamps into the ISO time format.

Although wallclock times are stored in db without a timezone, it still
gets sent back to the frontend with a Z at the end. It may be necessary
for the frontend to remove the Z to use the wallclock time as intended.