Make userId optional in createProject input
When creating a project, the default strategy will be to utilize the
userId stored in the JWT cookie to populate the userId field of the
project. However, createProject will take an optional userId in its
input, in which case the request will succeed only if the userId
matches the JWT cookie, or if the request is made by an admin.