Add wrapper to catch database errors
When the database runs into an error, Knex throws an error with the
original postgres error message. Apollo-Server by default passes this
message in its response, which is not best practice. This wrapper is
intended to wrap around all knex functions to ensure that any database
errors are logged to console, then replaced with a generic error when
it is sent back to the requester.