Move enum constraints for `templateName` from db to TypeScript
An enum constraint for `templateName` on the db would be hard to change
during migrations (see knex/knex#1699). Thus, it makes sense to offload
the enum constraint to TS and ensure we only use the TS enum instead of
the string literal in our code to achieve the same safety.