Reorganize email template classes
Previously, an abstract class was created because it was thought that
the 'send email' method would be common across all templates. However,
some templates will need to be sent to one user, while other templates
need to be sent to a group of users. Thus, each template needs to have
its own `dispatch` method responsible for sending out the emails to
whichever recipient(s) the template is intended for.